Hey, parsing is super robust. Does not handle shading state change (e.g. shading coeff or which mtl to use). That is coming up next. For now just want to get some triangles on the screen

+++ scop_xx/inc/vector.h
@@ -6,7 +6,7 @@
 /*   By: dgaitsgo <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/03/23 17:31:14 by dgaitsgo          #+#    #+#             */
-/*   Updated: 2017/03/29 17:20:44 by dgaitsgo         ###   ########.fr       */
+/*   Updated: 2017/04/03 22:04:17 by dgaitsgo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,23 +16,23 @@
 
 typedef struct		s_vector2
 {
-	double			x;
-	double			y;
+	float			x;
+	float			y;
 }					t_vector2;
 
 typedef struct		s_vector3
 {
-	double			x;
-	double			y;
-	double			z;
+	float			x;
+	float			y;
+	float			z;
 }					t_vector3;
 
 typedef struct		s_vector
 {
-	double			x;
-	double			y;
-	double			z;
-	double			w;
+	float			x;
+	float			y;
+	float			z;
+	float			w;
 }					t_vector;
 
 typedef struct
@@ -53,33 +53,33 @@ t_vector			get_up_vector(t_matrix m);
 t_vector			get_backward_vector(t_matrix m);
 t_vector			get_translation_vector(t_matrix m);
 t_vector			get_scale_vector(t_matrix m);
-t_vector			new_point(double x, double y, double z, double w);
-t_vector			new_vector(double a, double b, double c);
+t_vector			new_point(float x, float y, float z, float w);
+t_vector			new_vector(float a, float b, float c);
 void				copy_vector(t_vector *dst, t_vector *src);
-void				set_vector(t_vector *a, double x, double y, double z);
+void				set_vector(t_vector *a, float x, float y, float z);
 int					compare_vector(t_vector a, t_vector b);
-t_vector			vector_pow(t_vector a, double n);
+t_vector			vector_pow(t_vector a, float n);
 t_vector			vector_inverse(t_vector a);
-t_vector			vector_const_divide(t_vector a, double b);
+t_vector			vector_const_divide(t_vector a, float b);
 t_vector			vector_divide(t_vector a, t_vector b);
-t_vector			*new_vector_ptr(double x, double y, double z);
+t_vector			*new_vector_ptr(float x, float y, float z);
 t_vector			vector_cross(t_vector a, t_vector b);
-t_vector			vector_scale(t_vector a, double c);
-t_vector			vector_inverse_scale(t_vector a, double c);
+t_vector			vector_scale(t_vector a, float c);
+t_vector			vector_inverse_scale(t_vector a, float c);
 t_vector			vector_add(t_vector a, t_vector b);
-t_vector			vector_const_add(t_vector a, double b);
+t_vector			vector_const_add(t_vector a, float b);
 t_vector			vector_subtract(t_vector a, t_vector b);
 t_vector			vector_multiply(t_vector a, t_vector b);
 t_vector			unit_vector(t_vector a);
-double				vector_length(t_vector a);
+float				vector_length(t_vector a);
 int					compare_vector(t_vector a, t_vector b);
-double				dot_product(t_vector a, t_vector b);
+float				dot_product(t_vector a, t_vector b);
 t_vector			matrix_transform_point(t_vector *pnt, t_matrix m);
 t_vector			matrix_transform_raster_point(t_vector *pnt, t_matrix m);
 t_vector			matrix_transform_direction(t_vector *dir, t_matrix m);
 t_vector			matrix_transform_normal(t_vector *normal, t_matrix m);
-void				rotate_x(t_matrix rx, double x);
-void				rotate_y(t_matrix ry, double y);
-void				rotate_z(t_matrix rz, double z);
+void				rotate_x(t_matrix rx, float x);
+void				rotate_y(t_matrix ry, float y);
+void				rotate_z(t_matrix rz, float z);
 
 #endif
